+++ ChangeLog.md
@@ -4,3 +4,9 @@
 
 - First release
 - Statistics for received, sent, clapped on unipos
+
+## 1.0.1 - 2018-05-30
+
+- Use ECMAScript 2017 syntax `async/await`
+- Use built-int `MutationObserver` and fetch API
+- Drop support for old browser